+++ include/rotor.hpp
@@ -13,6 +13,7 @@
 #include "rotor/actor_base.h"
 #include "rotor/address.hpp"
 #include "rotor/message.h"
+#include "rotor/registry.h"
 #include "rotor/supervisor.h"
 #include "rotor/system_context.h"