+++ include/rotor/behavior.h
@@ -7,6 +7,7 @@
 
 #include "address.hpp"
 #include <system_error>
+#include <unordered_set>
 
 namespace rotor {
 
@@ -113,8 +114,27 @@ struct supervisor_behavior_t : public actor_behavior_t {
      * and forwared to the system context */
     virtual void on_shutdown_fail(const address_ptr_t &address, const std::error_code &ec) noexcept;
 
-    /** \brief reaction on child initialization failure. By default the child is asked for shut down */
-    virtual void on_init_fail(const address_ptr_t &address, const std::error_code &ec) noexcept;
+    virtual void on_start_init() noexcept override;
+
+    /** \brief supervisor behaviour on child-actor initialization result
+     *
+     * - if child-initialization failed and a supervisor is in INITIALIZING phase,
+     * then the reaction defined by `supervisor_policy`, i.e. either shutdown
+     * self (and all children) or just failing child;
+     * - if child successfully initialized, then send start message to it;
+     * - if there are no more initializing children, then finalize actor initialization
+
+    */
+    virtual void on_init(const address_ptr_t &address, const std::error_code &ec) noexcept;
+
+    /** \brief records child-actor address if it was created during supervsisor initialization */
+    virtual void on_create_child(const address_ptr_t &address) noexcept;
+
+    /** \brief type for keeping list of initializing actors (during supervisor inititalization) */
+    using initializing_actors_t = std::unordered_set<address_ptr_t>;
+
+    /** \brief list of initializing actors (during supervisor inititalization) */
+    initializing_actors_t initializing_actors;
 };
 
 } // namespace rotor