+++ include/rotor/supervisor_config.h
@@ -7,6 +7,7 @@
 //
 
 #include <boost/date_time/posix_time/posix_time.hpp>
+#include "policy.h"
 
 namespace rotor {
 
@@ -18,6 +19,9 @@ namespace pt = boost::posix_time;
 struct supervisor_config_t {
     /** \brief how much time is allowed to spend in shutdown for children actor */
     pt::time_duration shutdown_timeout;
+
+    /** \brief how to behave if child-actor fails */
+    supervisor_policy_t policy = supervisor_policy_t::shutdown_self;
 };
 
 } // namespace rotor